make transport buttons and shuttle controller work with MIDI CC's
git-svn-id: svn://localhost/ardour2/trunk@1477 d708f5d6-7413-0410-9779-e7cbd77b26cf

+++ libs/ardour/ardour/session.h
@@ -582,7 +582,7 @@ class Session : public PBD::StatefulDestructible
 
 	struct cleanup_report {
 	    vector<string> paths;
-	    int32_t space;
+	    int64_t space;
 	};
 
 	int  cleanup_sources (cleanup_report&);
